userExtractor
4.22

+++ osa4/bloglist/controllers/blogs.js
@@ -1,4 +1,4 @@
-const notesRouter = require('express').Router()
+const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
@@ -6,23 +6,23 @@ const middleware = require("../utils/middleware")
 
 
 
-notesRouter.get('/', async(request, response) => {
+blogsRouter.get('/', async(request, response) => {
     const blogs = await Blog
         .find({}).populate('user', {name:1, username:1} )
 
     response.json(blogs.map(blog => blog.toJSON()))
   })
   
-notesRouter.post('/', async(request, response) => {
+blogsRouter.post('/', middleware.userExtractor, async(request, response) => {
     const body = request.body
+    const user = request.user
     const token = request.token
     const decodedToken = jwt.verify(token, process.env.SECRET)
+
     if (!token || !decodedToken.id) {
         return response.status(401).json({ error: 'token missing or invalid' })
       }
-    const user = await User.findById(decodedToken.id)
-
-
+    
     const blog = new Blog({
         title: body.title,
         author: body.author,
@@ -41,20 +41,34 @@ notesRouter.post('/', async(request, response) => {
     // ei toiminut ilman {validateModifiedOnly: true }
     await user.save({validateModifiedOnly: true })
 
-    response.json(savedBlog.toJSON())
+    const populatedBlog = await savedBlog.populate('user',{ username: 1, name:1 })
+
+
+    response.status(200).json(populatedBlog.toJSON())
   })
 
 
-  notesRouter.delete('/:id', async (request, response, next) => {
-      try{
+  blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+    const blog = await Blog.findById(request.params.id)
+    const token = request.token
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    const user = request.user
+    
+    if (!token || !decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    if ( blog.user.toString() === user.id.toString() ){
         await Blog.findByIdAndRemove(request.params.id)
         response.status(204).end()
-      } catch(exception){
-        next(exception)
-      }
+    } else {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+      
   })
 
-  notesRouter.put('/:id', async (request, response) => {
+  blogsRouter.put('/:id', async (request, response) => {
 
     const body = request.body
     const blog = {
@@ -73,4 +87,4 @@ notesRouter.post('/', async(request, response) => {
 
   })
 
-module.exports = notesRouter
+module.exports = blogsRouter
